+++ organic-search-analytics/inc/code/mysql.php
@@ -14,6 +14,8 @@
 	
 		const DB_TABLE_SEARCH_ANALYTICS = 'search_analytics';
 		const DB_TABLE_SETTINGS = 'settings';
+		const DB_TABLE_SAVED_REPORTS = 'report_saved';
+		const DB_TABLE_SAVED_REPORT_CATEGORIES = 'report_saved_categories';
 		
 		
 		/**
@@ -33,17 +35,26 @@
 		 *
 		 *  @param $table     String   Table name
 		 *  @param $valueString     String   A value string to insert into the database. "field1"="value1", "field2"="value2"
+		 *  @param $returnObj     Boool   Default: False.  If true return database object upon success.  If false, return bool upon success.
 		 *
-		 *  @returns   Mixed   true(bool) upon success,
+		 *  @returns   Mixed   true(bool) or database object (depending on $returnObj) upon success,
 		 *             else returns error messsage upon failure
 		 */
-		public function qryDBinsert($table,$valueString) {
-			$GLOBALS['db']->query("INSERT INTO $table VALUES ($valueString)");
-		
-			if(mysqli_error($GLOBALS['db'])) { printf("Errormessage: %s\n", mysqli_error($GLOBALS['db'])); } else { return true; }
+		public function qryDBinsert( $table, $valueString, $returnObj = false ) {
+			$mysqliObj = $GLOBALS['db']->query("INSERT INTO $table VALUES ($valueString)");
+
+			if(mysqli_error($GLOBALS['db'])) {
+				printf("Errormessage: %s\n", mysqli_error($GLOBALS['db']));
+			} else {
+				if( $returnObj ) {
+					return $GLOBALS['db'];
+				} else {
+					return true;
+				}
+			}
 		}
-		
-		
+
+
 		/**
 		 *  Update record in Database
 		 *